requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
numpy==1.24.4
pillow==11.2.1
propcache==0.3.1
psutil==7.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pyparsing==3.2.3
python-multipart==0.0.20
rdkit-pypi==2022.9.5
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
torch==2.7.0
torch-geometric==2.6.1
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
